refactor(Configuration):
 - Refactored functions and variables into hooks for the configuration page
